Type casting for ID type to guarantee it is and integer
During the creation of a new api_key.py I miss typed the variable ID, that contains the ID of my chat which should be an INTEGER and not a STRING(this makes the bot only answer to my telegram user).
To avoid that mistake in the future I added a new type casting.